STOCKS-63: Unify code for entity deletion

+++ android-client/app/src/main/java/de/njsm/stocks/android/repo/CrashLogRepository.java
@@ -44,6 +44,9 @@ public class CrashLogRepository {
     }
 
     public void delete(CrashLog t) {
+        if (t.getFile() == null)
+            return;
+
         boolean result = t.getFile().delete();
         if (result) {
             LOG.i("Removed crash log " + t.getName());
